+++ test/module/parse-tokens.test.ts
@@ -11,8 +11,8 @@ describe("Parse-Tokens", () => {
 	describe("CompilableParseToken", () => {
 		// Example class for testing purposes
 		class ExampleToken extends CompilableParseToken {
-			semanticAnalysis(): void { }
-			translateCtxAndChildren(): Array<string> { return []; }
+			protected semanticAnalysis(): void { }
+			protected translateCtxAndChildren(): Array<any> { return []; }
 		}
 
 		describe("sourceCode", () => {